+++ e2e/nestjs/src/app.controller.spec.ts
@@ -1,22 +1,31 @@
-import { Test, TestingModule } from '@nestjs/testing'
-import { AppController } from './app.controller'
-import { AppService } from './app.service'
+import { Test, TestingModule } from '@nestjs/testing';
+import { AppController } from './app.controller';
+import { AppService } from './app.service';
+import { HttpModule } from '@nestjs/axios';
 
 describe('AppController', () => {
-	let appController: AppController
+  let appController: AppController;
 
-	beforeEach(async () => {
-		const app: TestingModule = await Test.createTestingModule({
-			controllers: [AppController],
-			providers: [AppService],
-		}).compile()
+  beforeEach(async () => {
+    const app: TestingModule = await Test.createTestingModule({
+      imports: [HttpModule],
+      controllers: [AppController],
+      providers: [AppService],
+    }).compile();
 
-		appController = app.get<AppController>(AppController)
-	})
+    appController = app.get<AppController>(AppController);
+  });
 
-	describe('root', () => {
-		it('should return "Hello World!"', () => {
-			expect(appController.getHello()).toBe('Hello World!')
-		})
-	})
-})
+  describe('root', () => {
+    it('should return "Hello World!"', () => {
+      expect(appController.getHello()).toBe('Hello World!');
+    });
+  });
+  describe('error', () => {
+    it('should throw error', () => {
+      expect(appController.getError()).toThrow(
+        new Error('a random error occurred!'),
+      );
+    });
+  });
+});