+++ e2e/nextjs/src/app/layout.tsx
@@ -37,7 +37,6 @@ export default function RootLayout({
 					canvasClearWebGLBuffer: false,
 				}}
 				backendUrl={CONSTANTS.NEXT_PUBLIC_HIGHLIGHT_BACKEND_URL}
-				scriptUrl={CONSTANTS.NEXT_PUBLIC_HIGHLIGHT_SCRIPT_URL}
 			/>
 
 			<html lang="en" data-layout>