+++ frontend/src/pages/LogsPage/LogsTable/LogsTable.tsx
@@ -31,10 +31,10 @@ import {
 	flexRender,
 	getCoreRowModel,
 	getExpandedRowModel,
+	Row,
 	useReactTable,
 } from '@tanstack/react-table'
 import { useVirtualizer } from '@tanstack/react-virtual'
-import { isEqual } from 'lodash'
 import React, { Key, useEffect, useMemo, useRef, useState } from 'react'
 
 import {
@@ -389,7 +389,7 @@ export const IconCollapsed: React.FC = () => (
 )
 
 type LogsTableRowProps = {
-	row: any
+	row: Row<LogEdgeWithResources>
 	rowVirtualizer: any
 	expanded: boolean
 	virtualRowKey: Key
@@ -397,86 +397,77 @@ type LogsTableRowProps = {
 	gridColumns: string[]
 }
 
-const LogsTableRow = React.memo<LogsTableRowProps>(
-	({
-		row,
-		rowVirtualizer,
-		expanded,
-		virtualRowKey,
-		queryParts,
-		gridColumns,
-	}) => {
-		const attributesRow = (row: any) => {
-			const log = row.original.node
-			const rowExpanded = row.getIsExpanded()
+const LogsTableRow: React.FC<LogsTableRowProps> = ({
+	row,
+	rowVirtualizer,
+	expanded,
+	virtualRowKey,
+	queryParts,
+	gridColumns,
+}) => {
+	const attributesRow = (row: LogsTableRowProps['row']) => {
+		const log = row.original.node
+		const rowExpanded = row.getIsExpanded()
 
-			const matchedAttributes = findMatchingAttributes(queryParts, {
-				...log.logAttributes,
-				environment: log.environment,
-				level: log.level,
-				message: log.message,
-				secure_session_id: log.secureSessionID,
-				service_name: log.serviceName,
-				service_version: log.serviceVersion,
-				source: log.source,
-				span_id: log.spanID,
-				trace_id: log.traceID,
-			})
-
-			return (
-				<Table.Row
-					selected={expanded}
-					className={styles.attributesRow}
-					gridColumns={['32px', '1fr']}
-				>
-					{rowExpanded && (
-						<>
-							<Table.Cell py="4" />
-							<Table.Cell py="4" borderTop="dividerWeak">
-								<LogDetails
-									matchedAttributes={matchedAttributes}
-									row={row}
-									queryParts={queryParts}
-								/>
-							</Table.Cell>
-						</>
-					)}
-				</Table.Row>
-			)
-		}
+		const matchedAttributes = findMatchingAttributes(queryParts, {
+			...log.logAttributes,
+			environment: log.environment,
+			level: log.level,
+			message: log.message,
+			secure_session_id: log.secureSessionID,
+			service_name: log.serviceName,
+			service_version: log.serviceVersion,
+			source: log.source,
+			span_id: log.spanID,
+			trace_id: log.traceID,
+		})
 
 		return (
-			<div
-				key={virtualRowKey}
-				data-index={virtualRowKey}
-				ref={rowVirtualizer.measureElement}
+			<Table.Row
+				selected={expanded}
+				className={styles.attributesRow}
+				gridColumns={['32px', '1fr']}
 			>
-				<Table.Row
-					gridColumns={gridColumns}
-					onClick={row.getToggleExpandedHandler()}
-					selected={expanded}
-					className={styles.dataRow}
-				>
-					{row.getVisibleCells().map((cell: any) => {
-						return (
-							<React.Fragment key={cell.column.id}>
-								{flexRender(
-									cell.column.columnDef.cell,
-									cell.getContext(),
-								)}
-							</React.Fragment>
-						)
-					})}
-				</Table.Row>
-				{attributesRow(row)}
-			</div>
+				{rowExpanded && (
+					<>
+						<Table.Cell py="4" />
+						<Table.Cell py="4" borderTop="dividerWeak">
+							<LogDetails
+								matchedAttributes={matchedAttributes}
+								row={row}
+								queryParts={queryParts}
+							/>
+						</Table.Cell>
+					</>
+				)}
+			</Table.Row>
 		)
-	},
-	(prevProps, nextProps) => {
-		return (
-			prevProps.expanded === nextProps.expanded &&
-			prevProps.virtualRowKey === nextProps.virtualRowKey &&
-			isEqual(prevProps.gridColumns, nextProps.gridColumns)
-		)
-	},
-)
+	}
+
+	return (
+		<div
+			key={virtualRowKey}
+			data-index={virtualRowKey}
+			ref={rowVirtualizer.measureElement}
+		>
+			<Table.Row
+				gridColumns={gridColumns}
+				onClick={row.getToggleExpandedHandler()}
+				selected={expanded}
+				className={styles.dataRow}
+			>
+				{row.getVisibleCells().map((cell: any) => {
+					return (
+						<React.Fragment key={cell.column.id}>
+							{flexRender(
+								cell.column.columnDef.cell,
+								cell.getContext(),
+							)}
+						</React.Fragment>
+					)
+				})}
+			</Table.Row>
+			{attributesRow(row)}
+		</div>
+	)
+}